refactor: adjust fee input
